added testing

+++ src/jair_turtlesim/follower/pid.py
@@ -1,16 +1,17 @@
-def limit(_value: int, _min: int, _max: int) -> int:
+def limit(_value: float, _min: float, _max: float) -> float:
     _value = max(_min, _value)
     _value = min(_max, _value)
     return _value
 
 
+# pylint: disable=too-few-public-methods
 class PID:
     def __init__(self, p: float = 0.8, i: float = 0.01, d: float = 0.05):
-        self._output_limit = 4
+        self._output_limit = 4.0
         self._p = p
         self._d = d
         self._i = i
-        self._i_growing = 0
+        self._i_growing = 0.0
         self._last_error = 0.0
 
     def calc(self, error: float) -> float:
